Refactor `bson2` package (#4105)
