+++ internal/bson2/document.go
@@ -18,12 +18,8 @@ import (
 	"bytes"
 	"encoding/binary"
 	"errors"
-	"fmt"
 	"log/slog"
 	"math"
-	"time"
-
-	"github.com/cristalhq/bson/bsonproto"
 
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/iterator"
@@ -140,8 +136,14 @@ func (doc *Document) Get(name string) any {
 
 // add adds a new field to the Document.
 func (doc *Document) add(name string, value any) error {
-	if err := validBSON(value); err != nil {
+	if err := validBSONType(value); err != nil {
 		return lazyerrors.Errorf("%q: %w", name, err)
+	}
+
+	if f, ok := value.(float64); ok {
+		if noNaN && math.IsNaN(f) {
+			return lazyerrors.New("invalid float64 value NaN")
+		}
 	}
 
 	doc.fields = append(doc.fields, field{
@@ -181,143 +183,6 @@ func (doc *Document) Encode() (RawDocument, error) {
 // LogValue implements slog.LogValuer interface.
 func (doc *Document) LogValue() slog.Value {
 	return slogValue(doc)
-}
-
-// encodeField encodes document field.
-//
-// It panics if v is not a valid type.
-func encodeField(buf *bytes.Buffer, name string, v any) error {
-	switch v := v.(type) {
-	case *Document:
-		if err := buf.WriteByte(byte(tagDocument)); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b := make([]byte, bsonproto.SizeCString(name))
-		bsonproto.EncodeCString(b, name)
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b, err := v.Encode()
-		if err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-	case RawDocument:
-		if err := buf.WriteByte(byte(tagDocument)); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b := make([]byte, bsonproto.SizeCString(name))
-		bsonproto.EncodeCString(b, name)
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		if _, err := buf.Write(v); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-	case *Array:
-		if err := buf.WriteByte(byte(tagArray)); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b := make([]byte, bsonproto.SizeCString(name))
-		bsonproto.EncodeCString(b, name)
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b, err := v.Encode()
-		if err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-	case RawArray:
-		if err := buf.WriteByte(byte(tagArray)); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		b := make([]byte, bsonproto.SizeCString(name))
-		bsonproto.EncodeCString(b, name)
-
-		if _, err := buf.Write(b); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-		if _, err := buf.Write(v); err != nil {
-			return lazyerrors.Error(err)
-		}
-
-	default:
-		return encodeScalarField(buf, name, v)
-	}
-
-	return nil
-}
-
-// encodeScalarField encodes scalar document field.
-//
-// It panics if v is not a scalar value.
-func encodeScalarField(buf *bytes.Buffer, name string, v any) error {
-	switch v := v.(type) {
-	case float64:
-		if noNaN && math.IsNaN(v) {
-			return lazyerrors.Errorf("got NaN value")
-		}
-		buf.WriteByte(byte(tagFloat64))
-	case string:
-		buf.WriteByte(byte(tagString))
-	case Binary:
-		buf.WriteByte(byte(tagBinary))
-	case ObjectID:
-		buf.WriteByte(byte(tagObjectID))
-	case bool:
-		buf.WriteByte(byte(tagBool))
-	case time.Time:
-		buf.WriteByte(byte(tagTime))
-	case NullType:
-		buf.WriteByte(byte(tagNull))
-	case Regex:
-		buf.WriteByte(byte(tagRegex))
-	case int32:
-		buf.WriteByte(byte(tagInt32))
-	case Timestamp:
-		buf.WriteByte(byte(tagTimestamp))
-	case int64:
-		buf.WriteByte(byte(tagInt64))
-	default:
-		panic(fmt.Sprintf("invalid BSON type %T", v))
-	}
-
-	b := make([]byte, bsonproto.SizeCString(name))
-	bsonproto.EncodeCString(b, name)
-
-	if _, err := buf.Write(b); err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	b = make([]byte, bsonproto.SizeAny(v))
-	bsonproto.EncodeAny(b, v)
-
-	if _, err := buf.Write(b); err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	return nil
 }
 
 // check interfaces
